+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Add initial `gke` package.
 - Add function `AddAuthForCluster` for having local config.
 - Add function `StreamContainerLog` for delegating a containers remote log to local files.
+- Add new type `PodRun` for executing a Pod for testing.